engine updates and other optimizations

+++ include/Food.h
@@ -22,11 +22,11 @@ private:
 	// in glm::ivec2, x represents the horizontal position and y represents the vertical position
 	static const int STARTING_X = 25, STARTING_Y = 10;
 
-	std::uniform_int_distribution<size_t> getRandomX_;
-	std::uniform_int_distribution<size_t> getRandomY_;
+	std::uniform_int_distribution<int> getRandomX_;
+	std::uniform_int_distribution<int> getRandomY_;
 
 	Evolve::TextureData foodTextures_[20];
-	int currentFoodIndex_ = 0;
+	size_t currentFoodIndex_ = 0;
 
 	std::uniform_int_distribution<size_t> getRandomFood_;
 };
